Don't run PostgreSQL tests on non-Linux
