+++ database/pubsub_test.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 package database_test
 
 import (
@@ -31,24 +33,6 @@ func TestPubsub(t *testing.T) {
 		data := "testing"
 		ch := make(chan []byte)
 		cancelFunc, err = pubsub.Subscribe(event, func(ctx context.Context, message []byte) {
-			ch <- message
-		})
-		require.NoError(t, err)
-		defer cancelFunc()
-		go func() {
-			err = pubsub.Publish(event, []byte(data))
-			require.NoError(t, err)
-		}()
-		message := <-ch
-		assert.Equal(t, string(message), data)
-	})
-
-	t.Run("Memory", func(t *testing.T) {
-		pubsub := database.NewPubsubInMemory()
-		event := "test"
-		data := "testing"
-		ch := make(chan []byte)
-		cancelFunc, err := pubsub.Subscribe(event, func(ctx context.Context, message []byte) {
 			ch <- message
 		})
 		require.NoError(t, err)
